airReserve/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
psycopg2==2.9.9
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.1.1
sqlparse==0.4.4
tzdata==2024.1
